Do not hardcore origin URLs
URLs should not be hardcoded in code, instead it needs to be specified on build time.

To accomplish this, a build-time environment file is required, which is done by adding `dotenv` as a dependency.

After merging this, copy env.build.txt to env.txt, then run yarn build and yarn data as usual.