Decouple fixModule from manifest fetching and fix mission translation etc.
The latest way of patching the game's code involved loading manifest.json first
before the game can be patched.

However, if the game loads first before the patching can occur, we're out of
luck. Because of this, fixModule needs to be run very early, before the game
even loads.

This commit decouples fixModule from the manifest loading, so it can always
load before Shiny. This comes with the cost of having to update the script
every new game update, however the relevant code needs to be run as early as
possible, and the only way we can ensure it is to incorporate it in the script
itself.

This commit also incorporates [this commit from the Chinese patch](https://github.com/biuuu/ShinyColors/commit/a0d9bc39324420b05ba0ba6921223834bb6a4e7b)
that fixes the request module to enable other translations.